rename src/node_modules to libs
