+++ src/br/phone/br-phone.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var StringMask = require('string-mask');
-var maskFactory = require('mask-factory');
+var maskFactory = require('../../libs/mask-factory');
 
 /**
  * FIXME: all numbers will have 9 digits after 2016.
