+++ src/ch/phone/ch-phone.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var StringMask = require('string-mask');
-var maskFactory = require('mask-factory');
+var maskFactory = require('../../libs/mask-factory');
 
 var phoneMask = new StringMask('+00 00 000 00 00');
